feat: Move getPublication function to with-cache module for better organization
