+++ apps/legal-gazette-public-web/src/app/(protected)/auglysingar/@sidebar/[id]/[version]/page.tsx
@@ -1,6 +1,6 @@
 import { PublicationSidebar } from '../../../../../../components/client-components/detailed-page/Sidebar/PublicationSidebar'
 import { GetAdvertPublicationVersionEnum } from '../../../../../../gen/fetch'
-import { getPublication } from '../../../../../../lib/cache/publication'
+import { getPublication } from '../../../../../../lib/with-cache/publication'
 
 export default async function AdvertPageSidebar({
   params,
